requirements.txt
pandas
numpy
jupyter
notebook
python-dotenv
openai